Moving room_index_data out of merc.h
